Refactor: 예외 시그니처 수정으로 인한 부가 수정
- Refs. #10
